Fix bad json response for courses
merged list of local and sis courses weren't in array form
